Move user validator methods
