+++ app/Controller/User.php
@@ -108,7 +108,7 @@ class User extends Base
     public function save()
     {
         $values = $this->request->getValues();
-        list($valid, $errors) = $this->user->validateCreation($values);
+        list($valid, $errors) = $this->userValidator->validateCreation($values);
 
         if ($valid) {
             $project_id = empty($values['project_id']) ? 0 : $values['project_id'];
@@ -329,7 +329,7 @@ class User extends Base
 
         if ($this->request->isPost()) {
             $values = $this->request->getValues();
-            list($valid, $errors) = $this->user->validatePasswordModification($values);
+            list($valid, $errors) = $this->userValidator->validatePasswordModification($values);
 
             if ($valid) {
                 if ($this->user->update($values)) {
@@ -371,7 +371,7 @@ class User extends Base
                 }
             }
 
-            list($valid, $errors) = $this->user->validateModification($values);
+            list($valid, $errors) = $this->userValidator->validateModification($values);
 
             if ($valid) {
                 if ($this->user->update($values)) {
@@ -409,7 +409,7 @@ class User extends Base
 
         if ($this->request->isPost()) {
             $values = $this->request->getValues() + array('disable_login_form' => 0, 'is_ldap_user' => 0);
-            list($valid, $errors) = $this->user->validateModification($values);
+            list($valid, $errors) = $this->userValidator->validateModification($values);
 
             if ($valid) {
                 if ($this->user->update($values)) {
